+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `EntityRef::has` to efficiently check for the presence of a component without borrowing it
 - `EntityBulider::new_cloneable` constructs an `EntityBuilder` that builds a `ReusableBuiltEntity`,
   which can be spawned repeatedly by reference.
+- `Satisfies` query combinator, which yields a `bool` without borrowing any components
 
 ### Changed
 - Added a niche to `Entity`, making `Option<Entity>` the same size as a bare `Entity`. As a